Address review comments. Export WaitForEndpoints to share between lib/update and lib/environ
