+++ lib/update/phase_kubernetes.go
@@ -18,7 +18,6 @@ package update
 
 import (
 	"context"
-	"time"
 
 	"github.com/gravitational/gravity/lib/constants"
 	"github.com/gravitational/gravity/lib/defaults"
@@ -26,9 +25,7 @@ import (
 	"github.com/gravitational/gravity/lib/kubernetes"
 	"github.com/gravitational/gravity/lib/loc"
 	"github.com/gravitational/gravity/lib/storage"
-	"github.com/gravitational/gravity/lib/utils"
 
-	"github.com/cenkalti/backoff"
 	"github.com/gravitational/rigging"
 	"github.com/gravitational/trace"
 	log "github.com/sirupsen/logrus"
@@ -36,7 +33,6 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	rbacv1beta1 "k8s.io/api/rbac/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	kubeapi "k8s.io/client-go/kubernetes"
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 )
@@ -215,7 +211,7 @@ func NewPhaseEndpoints(c FSMConfig, plan storage.OperationPlan, phase storage.Op
 // Execute waits for endpoints
 func (p *phaseEndpoints) Execute(ctx context.Context) error {
 	p.Infof("Wait for endpoints on %v.", p.Server)
-	err := waitForEndpoints(ctx, p.Client.CoreV1(), p.Server)
+	err := WaitForEndpoints(ctx, p.Client.CoreV1(), p.Server.KubeNodeID())
 	return trace.Wrap(err)
 }
 
@@ -387,87 +383,5 @@ func supportsTaints(gravityPackage loc.Locator) (supports bool, err error) {
 	}
 	return defaults.BaseTaintsVersion.Compare(*ver) <= 0, nil
 }
-
-func waitForEndpoints(ctx context.Context, client corev1.CoreV1Interface, server storage.Server) error {
-	node := server.KubeNodeID()
-	clusterLabels := labels.Set{"app": defaults.GravityClusterLabel}
-	kubednsLegacyLabels := labels.Set{"k8s-app": "kube-dns"}
-	kubednsLabels := labels.Set{"k8s-app": defaults.KubeDNSLabel}
-	matchesNode := matchesNode(node)
-	err := retry(ctx, func() error {
-		if (hasEndpoints(client, clusterLabels, existingEndpoint) == nil) &&
-			(hasEndpoints(client, kubednsLabels, matchesNode) == nil ||
-				hasEndpoints(client, kubednsLegacyLabels, matchesNode) == nil) {
-			return nil
-		}
-		return trace.NotFound("endpoints not ready")
-	}, defaults.EndpointsWaitTimeout)
-	return trace.Wrap(err)
-}
-
-func hasEndpoints(client corev1.CoreV1Interface, labels labels.Set, fn endpointMatchFn) error {
-	// TODO(dmitri): this is to workaround an issue with DNS service temporarily gone in 5.3.x
-	// See https://github.com/gravitational/gravity.e/issues/3866
-	// This is not necessary in version 5.4.x and up
-	services, err := client.Services(metav1.NamespaceSystem).List(
-		metav1.ListOptions{
-			LabelSelector: labels.String(),
-		},
-	)
-	if err != nil {
-		log.WithError(err).Warn("Failed to query services.")
-		return trace.Wrap(rigging.ConvertError(err), "failed to query services")
-	}
-	if len(services.Items) == 0 {
-		// Ignore endpoints for non-existing service (see comment above)
-		return nil
-	}
-	list, err := client.Endpoints(metav1.NamespaceSystem).List(
-		metav1.ListOptions{
-			LabelSelector: labels.String(),
-		},
-	)
-	if err != nil {
-		log.WithError(err).Warn("Failed to query endpoints.")
-		return trace.Wrap(rigging.ConvertError(err), "failed to query endpoints")
-	}
-	for _, endpoint := range list.Items {
-		for _, subset := range endpoint.Subsets {
-			for _, addr := range subset.Addresses {
-				log.WithField("addr", addr).Debug("Trying endpoint.")
-				if fn(addr) {
-					return nil
-				}
-			}
-		}
-	}
-	log.WithField("query", labels).Warn("No active endpoints found.")
-	return trace.NotFound("no active endpoints found for query %q", labels)
-}
-
-// matchesNode is a predicate that matches an endpoint address to the specified
-// node name
-func matchesNode(node string) endpointMatchFn {
-	return func(addr v1.EndpointAddress) bool {
-		// Abort if the node name is not populated.
-		// There is no need to wait for endpoints we cannot
-		// match to a node.
-		return addr.NodeName == nil || *addr.NodeName == node
-	}
-}
-
-// existingEndpoint is a trivial predicate that matches for any endpoint.
-func existingEndpoint(v1.EndpointAddress) bool {
-	return true
-}
-
-func retry(ctx context.Context, fn func() error, timeout time.Duration) error {
-	b := backoff.NewExponentialBackOff()
-	b.MaxElapsedTime = timeout
-	return trace.Wrap(utils.RetryWithInterval(ctx, b, fn))
-}
-
-// endpointMatchFn matches an endpoint address using custom criteria.
-type endpointMatchFn func(addr v1.EndpointAddress) bool
 
 type addTaint bool
